+++ app/javascript/packs/router/router.ts
@@ -1,7 +1,7 @@
 import Vue from 'vue/dist/vue.esm.js'
 import VueRouter from 'vue-router'
 import ArticlesContainer from '../container/ArticlesContainer.vue'
-import LoginContainer from '../container/login_container.vue'
+import LoginContainer from '../container/LoginContainer.vue'
 import UserRegistrationContainer from '../container/user_registration_container.vue'
 import CompleteUserRegistrationContainer from '../container/AfterUserRegistrationContainer.vue'
 import PostArticlesContainer from '../container/post_articles_container.vue'
